nos181: refactored onboarding to use NavigationStack instead of TabBar and moved each screen of onboarding flow into its own file
